backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
google-cloud-storage==2.14.0
google-cloud-dlp==3.15.0
google-cloud-aiplatform==1.70.0
google-cloud-secret-manager==2.18.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pdf2image==1.17.0
Pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
